requirements.txt
pyserial==3.5
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
